// grpc-server-product/src/proto/employees.ts
export interface CreateEmployeeDto {
  name: string;
  email: string;
  birthday?: Date | undefined;
  address?: string | undefined;
  allowance?: number | undefined;
  phone?: string | undefined;
  gender?: string | undefined;
}

export interface UpdateEmployeeDto {
  employeeId: string;
  name?: string | undefined;
  email?: string | undefined;
  birthday?: Date | undefined;
  address?: string | undefined;
  allowance?: number | undefined;
  phone?: string | undefined;
  gender?: string | undefined;
}

export interface SearchEmployeeDto {
  employeeId?: string | undefined;
  name?: string | undefined;
  email?: string | undefined;
}

export interface Employee {
  employeeId: string;
  name: string;
  email: string;
  birthday: Date | undefined;
  address: string;
  allowance: number;
  phone: string;
  gender: string;
  createdAt: Date | undefined;
  deletedAt: Date | undefined;
  updatedAt: Date | undefined;
}

export interface AllEmployees {
  employees: Employee[];
  length?: number | undefined;
}

export interface ResponseEmployee {
  code: number;
  message?: string | undefined;
  data: Employee | undefined;
}

export interface ResponseListEmployee {
  code: number;
  data: Employee[];
  message?: string | undefined;
}